# Efficient fresh name generation