--- a/plugins/cc/cctac.ml
+++ b/plugins/cc/cctac.ml
@@ -444,7 +444,7 @@ let cc_tactic depth additionnal_terms =
             let terms_to_complete = List.map (build_term_to_complete uf) (epsilons uf) in
             let hole = DAst.make @@ GHole (Evar_kinds.InternalHole, Misctypes.IntroAnonymous, None) in
             let pr_missing (c, missing) =
-              let c = Detyping.detype Detyping.Now ~lax:true false [] env sigma c in
+              let c = Detyping.detype Detyping.Now ~lax:true false Id.Set.empty env sigma c in
               let holes = List.init missing (fun _ -> hole) in
               Printer.pr_glob_constr_env env (DAst.make @@ GApp (c, holes))
             in