Use sets instead of lists in fresh name generation.

The generation of name was using O(n) check for freshness, which was a source of serious slowness observed in several parts, including tactics and printing. We use instead a check in O(log n) based on sets.

I'll launch a bench on pendulum to see if this has a more general impact.

This changes the API, so I don't think that this should be backported in 8.7.