@ppedrot 2017-09-10
Here are the timing diffs.

```
┌──────────────────────────┬─────────────────────────┬─────────────────────────────────────┬───────────────────────────────────────┬──────────────────────────┬─────────────────┐
│                          │      user time [s]      │             CPU cycles              │           CPU instructions            │  max resident mem [KB]   │   mem faults    │
│                          │                         │                                     │                                       │                          │                 │
│             package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │            NEW            OLD PDIFF   │     NEW     OLD  PDIFF   │ NEW OLD PDIFF   │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│   coq-mathcomp-ssreflect │   37.58   37.97 -1.03 % │  102135149123  102478293957 -0.33 % │   126734816509   126852097747 -0.09 % │  476876  476736  +0.03 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│                coq-sf-lf │   14.41   14.55 -0.96 % │   37580909794   37567187103 +0.04 % │    51334489765    51435398356 -0.20 % │  421592  421648  -0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│             coq-compcert │  806.74  814.37 -0.94 % │ 2233552109310 2260043838746 -1.17 % │  3396328232184  3395193759857 +0.03 % │ 1385704 1386080  -0.03 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│               coq-geocoq │ 2128.50 2142.62 -0.66 % │ 5925542825987 5957599934921 -0.54 % │  9971794642054  9990156908505 -0.18 % │ 1246836 1247784  -0.08 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│       coq-mathcomp-field │  451.39  453.35 -0.43 % │ 1256535094223 1262350394219 -0.46 % │  2102304740788  2101285461481 +0.05 % │  753424  753516  -0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│               coq-sf-plf │   40.59   40.76 -0.42 % │  111624797001  111630667060 -0.01 % │   144671700841   144801296630 -0.09 % │  510804  510852  -0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│                  coq-vst │ 1366.36 1371.75 -0.39 % │ 3794025816624 3807473671950 -0.35 % │  5689427817336  5689008381226 +0.01 % │ 2012128 1930116  +4.25 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│    coq-mathcomp-solvable │  189.22  189.78 -0.30 % │  524046063418  526086505936 -0.39 % │   763511422934   764065490080 -0.07 % │  713688  714420  -0.10 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│ coq-mathcomp-real_closed │  167.09  167.46 -0.22 % │  463222532372  464097641653 -0.19 % │   711845420511   711994626834 -0.02 % │  728784  729672  -0.12 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│                coq-flocq │   56.18   56.27 -0.16 % │  153627732244  153302796101 +0.21 % │   197948223963   197854022684 +0.05 % │  642964  643428  -0.07 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│         coq-fiat-parsers │  687.40  687.94 -0.08 % │ 1886719083982 1891642026829 -0.26 % │  3036200012023  3036545743927 -0.01 % │ 4132540 4132544  -0.00 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│           coq-coquelicot │   71.48   71.52 -0.06 % │  196490449258  196637443023 -0.07 % │   246774950809   246930790511 -0.06 % │  655892  656432  -0.08 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│            coq-fiat-core │  115.94  116.00 -0.05 % │  303922635337  304297153071 -0.12 % │   392021695429   392016811788 +0.00 % │  499408  499896  -0.10 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│   coq-mathcomp-character │  268.98  269.08 -0.04 % │  747159022562  748290719630 -0.15 % │  1097328465693  1097426840724 -0.01 % │  979512  979632  -0.01 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│     coq-mathcomp-algebra │  167.76  167.74 +0.01 % │  464942054569  465037764002 -0.02 % │   661224508191   660928676303 +0.04 % │  589212  589924  -0.12 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│          coq-fiat-crypto │ 2992.31 2991.94 +0.01 % │ 8288228816492 8283648128939 +0.06 % │ 13869845128411 13883730771487 -0.10 % │ 3239864 3223284  +0.51 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│                 coq-hott │  343.48  343.32 +0.05 % │  936354698669  935844342915 +0.05 % │  1585891399496  1585562370892 +0.02 % │  519936  520064  -0.02 % │   1   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│              coq-unimath │ 1153.96 1152.32 +0.14 % │ 3200897341847 3193474170728 +0.23 % │  5482341514341  5482539059119 -0.00 % │ 1062964 1062508  +0.04 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│               coq-sf-vfa │   17.75   17.71 +0.23 % │   48318378408   48460322385 -0.29 % │    65416827497    65518215335 -0.15 % │  530696  535840  -0.96 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│    coq-mathcomp-fingroup │   58.66   58.49 +0.29 % │  160932253333  161037916471 -0.07 % │   225753773233   225774035502 -0.01 % │  531216  529292  +0.36 % │   0   0  +nan % │
├──────────────────────────┼─────────────────────────┼─────────────────────────────────────┼───────────────────────────────────────┼──────────────────────────┼─────────────────┤
│   coq-mathcomp-odd_order │ 1377.41 1367.58 +0.72 % │ 3836293046000 3809540684054 +0.70 % │  6700886047479  6706360759266 -0.08 % │ 1084572 1206904 -10.14 % │   0   0  +nan % │
└──────────────────────────┴─────────────────────────┴─────────────────────────────────────┴───────────────────────────────────────┴──────────────────────────┴─────────────────┘


PDIFF = proportional difference between measurements done for the NEW and the OLD Coq version
      = (NEW_measurement - OLD_measurement) / OLD_measurement * 100%

NEW = 492811479081cd248d0cb07432ac269804c21378
OLD = b1fbec7e3945fe2965f4ba9f80c8c31b821dbce1
```